crawlers/program-membership-scanner/requirements.txt
python-magic==0.4.15
python-barcode==0.8.1
requests==2.18.4
